Session Length Tweaks
Include support for variable session lengths.  As long as the page is
refreshed the session will be regenerated as part of the verification
process, meaning that only a period of inactivity will result in session
expiration.

Support also included for the 'remember me' functionality which
significantly extends the storage time of the cookie, whether active or
otherwise.